(PC-20310)[API] refactor: use duplicate SubcategoryIdEnum for serializers retro compatibility
